Move all Firebase dependencies into studio-firebase package
