+++ package.json
@@ -104,7 +104,6 @@
     "eslint-plugin-prettier": "3.4.0",
     "eslint-plugin-react": "7.23.2",
     "eslint-plugin-react-hooks": "4.2.0",
-    "firebase": "9.0.0-beta.1",
     "fork-ts-checker-webpack-plugin": "6.2.9",
     "html-webpack-plugin": "5.3.1",
     "jest": "27.0.0-next.8",
