+++ packages/@foxglove/studio-firebase/jest.config.json
@@ -1,0 +1,9 @@
+{
+  "testRunner": "jest-circus/runner",
+  "testMatch": ["<rootDir>/src/**/*.test.ts"],
+  "transform": {
+    "\\.[jt]sx?$": ["babel-jest", { "rootMode": "upward" }]
+  },
+  "//": "Native find is slow because it does not exclude files: https://github.com/facebook/jest/pull/11264#issuecomment-825377579",
+  "haste": { "forceNodeFilesystemAPI": true }
+}
