+++ packages/@foxglove/studio-firebase/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@foxglove/studio-firebase",
+  "version": "0.0.0",
+  "license": "MPL-2.0",
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/foxglove/studio.git"
+  },
+  "author": {
+    "name": "Foxglove Technologies",
+    "email": "support@foxglove.dev"
+  },
+  "homepage": "https://foxglove.dev/",
+  "main": "./src/index.ts",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "scripts": {
+    "prepack": "tsc -b"
+  },
+  "devDependencies": {
+    "typescript": "4.2.4"
+  },
+  "dependencies": {
+    "@foxglove/log": "workspace:packages/log",
+    "@foxglove/studio-base": "workspace:app",
+    "firebase": "9.0.0-beta.1"
+  }
+}
