+++ packages/@foxglove/studio-firebase/src/context/FirebaseAppContext.tsx
