+++ packages/@foxglove/studio-firebase/src/index.ts
@@ -1,0 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+export { default as FirebaseAppProvider } from "./providers/FirebaseAppProvider";
+export { default as FirebaseAuthProvider } from "./providers/FirebaseAuthProvider";
+export { useFirebase } from "./context/FirebaseAppContext";
