+++ packages/@foxglove/studio-firebase/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@foxglove/tsconfig/tsconfig.base.json",
+  "include": ["./src/**/*"],
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "lib": ["dom", "dom.iterable", "es2020"],
+    "jsx": "react-jsx"
+  }
+}
