+++ web/src/Root.tsx
@@ -13,8 +13,8 @@ import {
   ThemeProvider,
   UserProfileLocalStorageProvider,
   StudioToastProvider,
-  FirebaseAppProvider,
 } from "@foxglove/studio-base";
+import { FirebaseAppProvider } from "@foxglove/studio-firebase";
 
 import AppConfigurationProvider from "./components/AppConfigurationProvider";
 import NoOpLayoutStorageProvider from "./components/NoOpLayoutStorageProvider";
